+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **feat**: add debounce for editor change event
 - **feat**: add stdoutThreshold configuration to control stdout output handling
 - **feat**: clear test case result
+- **feat**: add custom file system
+- **feat**: edit raw problem data
 
 ## 0.2.2